remove short_url and add in-tree URLencoder (#53)
This PR removes the short_url dependency as per issue #41.
This implementation is pretty much the same as in short_url except I've rewritten the enbase() function to be iterative instead of recursive.
The only functions of the class are enbase() and debase() since those were the only functions being used by fhost.

Co-authored-by: 7415963987456321 <hrs70@hi.is>
Reviewed-on: https://git.0x0.st/mia/0x0/pulls/53
Co-authored-by: mia <mia@0x0.st>
Co-committed-by: mia <mia@0x0.st>